+++ ouroboros-network-api/src/Ouroboros/Network/PeerSelection/RelayAccessPoint.hs
@@ -1,18 +1,22 @@
 {-# LANGUAGE BangPatterns      #-}
+{-# LANGUAGE LambdaCase        #-}
 {-# LANGUAGE NamedFieldPuns    #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE PatternSynonyms   #-}
+{-# LANGUAGE TypeApplications  #-}
 {-# LANGUAGE ViewPatterns      #-}
 
 module Ouroboros.Network.PeerSelection.RelayAccessPoint
   ( DomainAccessPoint (..)
   , RelayAccessPoint (.., RelayDomainAccessPoint)
+  , RelayAccessPointCoded (..)
   , IP.IP (..)
     -- * Socket type re-exports
   , Socket.PortNumber
   ) where
 
 import Control.DeepSeq (NFData (..))
+import Control.Monad (when)
 
 import Data.Aeson
 import Data.IP qualified as IP
@@ -21,6 +25,8 @@ import Data.Text qualified as Text
 import Data.Text.Encoding (decodeUtf8, encodeUtf8)
 import Text.Read (readMaybe)
 
+import Cardano.Binary (FromCBOR (..), ToCBOR (..))
+import Cardano.Binary qualified as Codec
 import Network.DNS qualified as DNS
 import Network.Socket qualified as Socket
 
@@ -52,6 +58,51 @@ instance ToJSON DomainAccessPoint where
 data RelayAccessPoint = RelayAccessDomain  !DNS.Domain !Socket.PortNumber
                       | RelayAccessAddress !IP.IP      !Socket.PortNumber
   deriving (Eq, Ord)
+
+newtype RelayAccessPointCoded = RelayAccessPointCoded { unRelayAccessPointCoded :: RelayAccessPoint }
+
+-- | These instances are used to serialize 'LedgerPeerSnapshot'
+-- consensus LocalStateQuery server which uses these instances
+-- for all its query responses. It appears they provide some improved
+-- debugging diagnostics over Serialize instances.
+instance ToCBOR RelayAccessPointCoded where
+  toCBOR (RelayAccessPointCoded rap) = case rap of
+    RelayAccessDomain domain port ->
+         Codec.encodeListLen 3
+      <> Codec.encodeWord8 0
+      <> serialise' port
+      <> toCBOR domain
+    RelayAccessAddress (IP.IPv4 ipv4) port ->
+         Codec.encodeListLen 3
+      <> Codec.encodeWord8 1
+      <> serialise' port
+      <> toCBOR (IP.fromIPv4 ipv4)
+    RelayAccessAddress (IP.IPv6 ip6) port ->
+         Codec.encodeListLen 3
+      <> Codec.encodeWord8 2
+      <> serialise' port
+      <> toCBOR (IP.fromIPv6 ip6)
+    where
+      serialise' = toCBOR . toInteger
+
+instance FromCBOR RelayAccessPointCoded where
+  fromCBOR = do
+    listLen <- Codec.decodeListLen
+    when (listLen /= 3) . fail $    "Unrecognized RelayAccessPoint list length "
+                                 <> show listLen
+    constructorTag <- Codec.decodeWord8
+    port <- fromInteger <$> fromCBOR @Integer
+    case constructorTag of
+      0 -> do
+        domain <- fromCBOR
+        return . RelayAccessPointCoded $ RelayAccessDomain domain port
+      1 -> do
+        ip4 <- IP.IPv4 . IP.toIPv4 <$> fromCBOR
+        return . RelayAccessPointCoded $ RelayAccessAddress ip4 port
+      2 -> do
+        ip6 <- IP.IPv6 . IP.toIPv6 <$> fromCBOR
+        return . RelayAccessPointCoded $ RelayAccessAddress ip6 port
+      _ -> fail $ "Unrecognized RelayAccessPoint tag: " <> show constructorTag
 
 instance Show RelayAccessPoint where
     show (RelayAccessDomain domain port) =